requirements.txt
absl-py==0.7.1
aiofiles==0.4.0
aiohttp==3.5.4
alembic==1.0.11
APScheduler==3.6.1
argh==0.26.2
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
async-generator==1.10
async-timeout==3.0.1
attrdict==2.0.1
attrs==19.1.0
azure-cosmos==3.1.2
base58==1.0.3
beautifulsoup4==4.8.0
blis==0.4.1
boto3==1.9.215
botocore==1.12.215
catalogue==0.0.8
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
ConfigArgParse==0.14.0
cryptography==2.7
cycler==0.10.0
cymem==2.0.3
cytoolz==0.10.0
decorator==4.4.0
docopt==0.6.2
docutils==0.15.2
es-core-news-sm==2.2.5
eth-abi==2.0.0
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.1.0
eth-utils==1.7.0
fakeredis==1.0.4
fbmessenger==6.0.0
Flask==1.1.1
Flask-Cors==3.0.8
future==0.17.1
gast==0.2.2
gevent==1.4.0
gitdb2==2.0.5
GitPython==3.0.2
greenlet==0.4.15
grpcio==1.23.0
h5py==2.9.0
hexbytes==0.2.0
httptools==0.0.13
humanfriendly==4.18
idna==2.8
importlib-metadata==0.23
ipfshttpclient==0.4.12
isodate==0.6.0
isort==4.3.16
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
Js2Py==0.66
jsonpickle==1.2
jsonschema==3.1.1
kafka-python==1.4.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
lru-dict==1.1.6
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mattermostwrapper==2.1
mccabe==0.6.1
mixpanel==4.4.0
mock==3.0.5
more-itertools==7.2.0
multiaddr==0.0.8
multidict==4.5.2
murmurhash==1.0.2
netaddr==0.7.19
networkx==2.3
numpy==1.17.0
packaging==19.1
parsimonious==0.8.1
pathtools==0.1.2
pika==1.0.1
pipwin==0.4.5
plac==1.1.3
pocketsphinx==0.1.15
preshed==3.0.2
prompt-toolkit==2.0.9
protobuf==3.9.1
psycopg2==2.8.4
PyAudio==0.2.11
pycparser==2.19
pycryptodome==3.9.0
pydot==1.4.1
pyfcm==1.4.7
pyjsparser==2.7.1
PyJWT==1.7.1
pykwalify==1.7.0
pylint==2.3.1
pymongo==3.9.0
pyparsing==2.4.2
pypiwin32==223
PyPrind==2.11.2
pyreadline==2.1
pyrsistent==0.15.4
PySocks==1.7.0
python-crfsuite==0.9.6
python-dateutil==2.8.0
python-editor==1.0.4
python-engineio==3.9.3
python-gnupg==0.4.5
python-socketio==4.3.1
python-telegram-bot==11.1.0
pytz==2019.2
pywin32==225
PyYAML==5.1.2
questionary==1.3.0
rasa==1.2.4
rasa-sdk==1.2.0
rasa-x==0.20.1
redis==3.3.8
requests==2.22.0
requests-toolbelt==0.9.1
rlp==1.1.0
robobrowser==0.5.3
rocketchat-API==0.6.34
ruamel.yaml==0.15.100
s3transfer==0.2.1
sanic==19.3.1
Sanic-Cors==0.9.9.post1
sanic-jwt==1.3.2
Sanic-Plugins-Framework==0.8.2
scikit-learn==0.20.4
scipy==1.3.1
setuptools-scm==3.3.3
simplejson==3.16.0
six==1.12.0
sklearn-crfsuite==0.3.6
slackclient==1.3.2
smmap2==2.0.5
sortedcontainers==2.1.0
soupsieve==1.9.3
spacy==2.2.3
SpeechRecognition==3.8.1
SQLAlchemy==1.3.7
srsly==0.2.0
tabulate==0.8.3
tensorboard==1.13.1
tensorflow==1.13.2
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminaltables==3.1.0
thinc==7.3.1
toolz==0.10.0
tqdm==4.35.0
twilio==6.29.4
typed-ast==1.3.1
tzlocal==2.0.0
urllib3==1.25.3
varint==1.0.2
wasabi==0.4.0
watchdog==0.9.0
wcwidth==0.1.7
web3==5.2.2
webexteamssdk==1.1.1
websocket-client==0.54.0
websockets==7.0
Werkzeug==0.15.5
wincertstore==0.2
wrapt==1.11.1
yarl==1.3.0
zipp==0.6.0
